Rename `remote.py` to `websockets.py`
